DLL matches (#121)
* Merging Anim DLL progress (and other DLL odds and ends) with latest

* Match: SnowHorn dll_496_func_84C (and some SnowHorn cleanup)

* Added scratch links for the remaining SnowHorn functions

* Match: dll_496_func_1D68 (thanks to Bloodangel!)

* Tumbleweed progress

* Tumbleweed matches

* Match: setuppoint DLL

* Match: area DLL

* Match: n_pausemenu DLL (with Shinx and Bloodangel's help!)

* Parameterised the Pause Menu's coordinates and colours (to help reposition the Duster icon box together with its text, etc.)

* Added controller button enum

* Found where the player DLL uses GplayStruct10 (stats like health, Dusters)

* Created n_pausemenu.h

* Renamed GplayStruct10 struct to PlayerStats

* Moved PlayerState struct to 210_player.h

* Merged in anim DLL strings

* More Tumbleweed DLL matches (thanks to Bloodangel!)

* Moved Tumbleweed data (still struggling with its rodata at the moment)

* Used the really handy NEW_PICMENU_ITEM macro on pausemenu, and renamed to just "pausemenu"

* Updated pausemenu syms to new format

* Removed controller.h button enum and used PR/os.h's ones instead

* Added names for DLLs 60 and 61

* Progress on mainmenu DLL

* Match: mainmenu DLL

* Added names to Menu enum, and added/edited menu DLL names (official names from SFA kiosk)

* Mainmenu cleanup - also filled in GplayStruct4 member names and renamed to GplayOptions

* Renamed pausemenu header

+++ include/functions.h
@@ -14,7 +14,7 @@ void dl_segment(Gfx **gdl, u32 segment, void *base);
 void stop_alSyn_thread();
 
 void func_800267A4(Object *obj);
-s32 func_80023D30(Object*,s32,f32,s32);
+s32 func_80023D30(Object*,s32,f32,u8);
 s32 func_80024108(Object*,f32,f32,s32);
 s32 func_80025F40(Object*,s32 *,s32 *,s32 *);
 
@@ -61,6 +61,9 @@ void draw_pause_screen_freeze_frame(Gfx** gdl);
 
 void func_80037A14(Gfx**, Mtx **, s32);
 
+/** Draws a 2D texture to screen, scaling it down horizontally (from its left edge) if
+  * widescreen is enabled in order to maintain aspect ratio 
+  */
 void func_8003825C(Gfx **gdl, Texture*, s32, s32, s32, s32, s32, s32);
 void func_800382AC(Gfx **gdl, Texture*, s32, s32, s32, s32, s32, s32);
